Api Rest Usuários, senha HashSaltPepper

+++ Backend/MercadoApi/Mercado.MOD/Produto.cs
@@ -4,7 +4,7 @@
     {
         public int? idProduto { get; set; }
         public string? nome { get; set; }
-        public int? valor { get; set; }
+        public float? valor { get; set; }
         public string? unidade { get; set; }
         public int? quantidade { get; set; }
